requirements.txt
shillelagh[gsheetsapi]
sqlalchemy
sqlalchemy.orm
requests
google_auth_oauthlib
google-auth
google-auth-httplib2
google-api-python-client